requirements.txt
-i https://pypi.org/simple

-e .

appdirs==1.4.4

beautifulsoup4==4.9.3

certifi==2020.12.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

colorama==0.4.4

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pafy==0.5.5

pyperclip==1.8.2

python-vlc==3.0.12118

requests==2.25.1

soupsieve==2.2.1; python_version >= '3.0'

urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

youtube-dl==2021.4.26

VLCYT==1.9
